Update LivewireDatatable.php (#79)
* Update LivewireDatatable.php

In getSortString() the default is to use a back tick " ` " for the order by field. This causes an error in postgresql which likes it to be ' " '. My fix is to test the enviroment variable, and if it is set to "pgsql" it will use ' " ' otherwise it will give " ` ".

* Update LivewireDatatable.php

Changed quotation style.